sound: use a 32-bit accumulator in the SFX resampler interior MAC
The polyphase up-sampling loop accumulated into an int64_t. That forced
the compiler to widen every 16-bit tap product to 64 bits, which blocks
the 16x16->32 packed multiply-add (pmaddwd on SSE2, smlal on NEON) that
the flat interior dot product is otherwise a textbook fit for.

The accumulator never needs 64 bits: the per-phase kernel tap magnitudes
sum to at most ~35700, so with 16-bit samples |acc| <= 35700 * 32768,
about 1.17e9, comfortably under INT_MAX. Switching the interior and edge
accumulators to int lets the loop vectorise (verified: pmaddwd is now
emitted where it previously was not) while producing byte-identical
output - an exhaustive old-vs-new sweep (4 input rates x 6 output rates x
both widths x 9 lengths x 6 signal types = 2592/2592 cases) stays
bit-for-bit identical, so mix/netplay/runahead determinism is preserved.
The down-sampling path (never taken for the usual source -> higher-rate
case) keeps its 64-bit accumulator.

Interior MAC ~2.6x faster in isolation; overall resampling drops a
further ~1.5x on a typical level load (213 -> 143 ms), on top of the
polyphase-bank speedup.

+++ client/snd_mem.c
@@ -570,13 +570,18 @@ void ResampleSfx (sfx_t *sfx, int inrate, int inwidth, byte *data)
 			const short	*kp   = s_sfx_bank[phase];
 			int		csum  = s_sfx_bankcsum[phase];
 			int		j, sample;
-			int64_t		acc = 0;
+			/* 32-bit accumulator: the per-phase kernel tap magnitudes sum
+			 * to at most ~35700, so |acc| <= 35700 * 32768 < 2^31 and cannot
+			 * overflow.  Keeping it int (not int64_t) lets the compiler emit
+			 * a 16x16->32 widening multiply-add (pmaddwd / smlal) for the
+			 * interior loop; results are identical to the 64-bit version. */
+			int		acc = 0;
 
 			if (base >= (SFX_RS_NZ - 1) && base <= inlength - (SFX_RS_NZ + 1))
 			{
 				const short *sp = src16 + base - (SFX_RS_NZ - 1);
 				for (j = 0; j < SFX_RS_TAPS; j++)
-					acc += (int64_t)kp[j] * sp[j];
+					acc += kp[j] * sp[j];
 			}
 			else
 			{
@@ -587,7 +592,7 @@ void ResampleSfx (sfx_t *sfx, int inrate, int inwidth, byte *data)
 						idx = 0;
 					else if (idx >= inlength)
 						idx = inlength - 1;
-					acc += (int64_t)kp[j] * src16[idx];
+					acc += kp[j] * src16[idx];
 				}
 			}
